added no posts yet component and extracted new post dialog component
