+++ src/views/dashboard/components/widgets/components/user-stats/user-stats.component.js
@@ -3,7 +3,22 @@ import { Typography, List, ListItem, Divider, Box } from '@mui/material';
 import { Stat, StyledLinearProgress } from './user-stats.styles';
 import { WidgetContainer } from '../../widgets.styles';
 
-const UserStats = ({ user }) => {
+const user = {
+    profilePicture: 'https://via.placeholder.com/150',
+    name: 'John Doe',
+    workouts: 150,
+    achievements: ['5k Runner', '100 Workouts', 'Marathon Finisher'],
+    goalProgress: 75, // This represents the percentage of the goal achieved.
+    recentActivities: [
+        'Completed a 10k run',
+        'Joined "Summer Fitness Challenge"',
+        'Achieved "5k Runner" badge',
+        'Posted a new workout routine',
+        'Shared a diet tip'
+    ]
+};
+
+const UserStats = () => {
     return (
         <WidgetContainer>
             <Typography sx={{ p: '20px' }} variant="h5" color="textPrimary">My Stats</Typography>
